requirements.txt
asgiref==3.2.10
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
Markdown==3.3.1
Pillow==8.0.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
